create a validator class for the face observation data.
