+++ Sources/SmileID/Classes/FaceDetector/FaceDetectorV2.swift
@@ -19,53 +19,7 @@ class FaceDetectorV2: NSObject {
     weak var selfieViewModel: SelfieViewModelV2?
     weak var viewDelegate: FaceDetectorDelegate?
 
-    // MARK: Publishers for Vision data
-    private var hasDetectedValidFace: Bool
-    private var faceDetectedState: FaceDetectionState {
-        didSet {
-            determineDirective()
-        }
-    }
-    private var faceGeometryState: FaceObservation<FaceGeometryModel> {
-        didSet {
-            processUpdatedFaceGeometry()
-        }
-    }
-    private var faceQualityState: FaceObservation<FaceQualityModel> {
-        didSet {
-            processUpdatedFaceQuality()
-        }
-    }
-    private var selfieQualityState: FaceObservation<SelfieQualityModel> {
-        didSet {
-            processUpdatedSelfieQuality()
-        }
-    }
-    private var isAcceptableBounds: FaceBoundsState {
-        didSet {
-            calculateDetectedFaceValidity()
-        }
-    }
-    private var isAcceptableFaceQuality: Bool {
-        didSet {
-            calculateDetectedFaceValidity()
-        }
-    }
-    private var isAcceptableSelfieQuality: Bool {
-        didSet {
-            calculateDetectedFaceValidity()
-        }
-    }
-
     override init() {
-        self.hasDetectedValidFace = false
-        self.faceDetectedState = .noFaceDetected
-        self.faceGeometryState = .faceNotFound
-        self.faceQualityState = .faceNotFound
-        self.selfieQualityState = .faceNotFound
-        self.isAcceptableBounds = .unknown
-        self.isAcceptableFaceQuality = false
-        self.isAcceptableSelfieQuality = false
         super.init()
         selfieQualityModel = createImageClassifier()
     }
@@ -145,7 +99,7 @@ class FaceDetectorV2: NSObject {
                 failed: failScore,
                 passed: passScore
             )
-            // selfieViewModel.perform(action: .selfieQualityObservationDetected(selfieQualityModel))
+            selfieViewModel.perform(action: .selfieQualityObservationDetected(selfieQualityModel))
         } catch {
             selfieViewModel.perform(action: .handleError(error))
         }
@@ -188,102 +142,6 @@ class FaceDetectorV2: NSObject {
     }
 }
 
-// MARK: Perform Checks
-extension FaceDetectorV2 {
-    func determineDirective() {
-        guard let viewModel = selfieViewModel else { return }
-        switch faceDetectedState {
-        case .faceDetected:
-            if hasDetectedValidFace {
-                viewModel.perform(action: .updateUserInstruction(nil))
-            } else if isAcceptableBounds == .detectedFaceTooSmall {
-                viewModel.perform(action: .updateUserInstruction(.moveCloser))
-            } else if isAcceptableBounds == .detectedFaceTooLarge {
-                viewModel.perform(action: .updateUserInstruction(.moveBack))
-            } else if isAcceptableBounds == .detectedFaceOffCentre {
-                viewModel.perform(action: .updateUserInstruction(.headInFrame))
-            } else if !isAcceptableSelfieQuality {
-                viewModel.perform(action: .updateUserInstruction(.goodLight))
-            }
-        case .noFaceDetected:
-            viewModel.perform(action: .updateUserInstruction(.headInFrame))
-        case .faceDetectionErrored:
-            viewModel.perform(action: .updateUserInstruction(nil))
-        }
-    }
-
-    func processUpdatedFaceGeometry() {
-        switch faceGeometryState {
-        case let .faceFound(faceGeometryModel):
-            let boundingBox = faceGeometryModel.boundingBox
-            updateAcceptableBounds(using: boundingBox)
-        //            if hasDetectedValidFace && selfieViewModel.selfieImage != nil && activeLiveness.currentTask != nil {
-        //                activeLiveness.processFaceGeometry(faceGeometryModel)
-        //            }
-        case .faceNotFound:
-            invalidateFaceGeometryState()
-        case let .errored(error):
-            print(error.localizedDescription)
-            invalidateFaceGeometryState()
-        }
-    }
-
-    func invalidateFaceGeometryState() {
-        // This is where we reset all the face geometry values.
-        isAcceptableBounds = .unknown
-    }
-
-    func updateAcceptableBounds(using boundingBox: CGRect) {
-        guard let selfieViewModel else { return }
-        let faceLayoutGuideFrame = selfieViewModel.faceLayoutGuideFrame
-        if boundingBox.width > 1.2 * faceLayoutGuideFrame.width {
-            isAcceptableBounds = .detectedFaceTooLarge
-        } else if boundingBox.width * 1.2 < faceLayoutGuideFrame.width {
-            isAcceptableBounds = .detectedFaceTooSmall
-        } else {
-            if abs(boundingBox.midX - faceLayoutGuideFrame.midX) > 50 {
-                isAcceptableBounds = .detectedFaceOffCentre
-            } else if abs(boundingBox.midY - faceLayoutGuideFrame.midY) > 50 {
-                isAcceptableBounds = .detectedFaceOffCentre
-            } else {
-                isAcceptableBounds = .detectedFaceAppropriateSizeAndPosition
-            }
-        }
-    }
-
-    func processUpdatedFaceQuality() {
-        switch faceQualityState {
-        case let .faceFound(faceQualityModel):
-            // Check acceptable range here.
-            isAcceptableFaceQuality = faceQualityModel.quality > 0.2
-        case .faceNotFound:
-            isAcceptableFaceQuality = false
-        case let .errored(error):
-            print(error.localizedDescription)
-            isAcceptableFaceQuality = false
-        }
-    }
-
-    func processUpdatedSelfieQuality() {
-        switch selfieQualityState {
-        case let .faceFound(selfieQualityModel):
-            // Check acceptable range here.
-            isAcceptableSelfieQuality = selfieQualityModel.passed > 0.5
-        case .faceNotFound:
-            isAcceptableSelfieQuality = false
-        case let .errored(error):
-            print(error.localizedDescription)
-            isAcceptableSelfieQuality = false
-        }
-    }
-
-    func calculateDetectedFaceValidity() {
-        hasDetectedValidFace =
-            isAcceptableBounds == .detectedFaceAppropriateSizeAndPosition && isAcceptableFaceQuality
-            && isAcceptableSelfieQuality
-    }
-}
-
 // MARK: - Private methods
 extension FaceDetectorV2 {
     func detectedFaceRectangles(request: VNRequest, error: Error?) {
@@ -308,8 +166,7 @@ extension FaceDetectorV2 {
                 pitch: result.pitch ?? 0.0,
                 direction: faceDirection(faceObservation: result)
             )
-            self.faceDetectedState = .faceDetected
-            self.faceGeometryState = .faceFound(faceObservationModel)
+            selfieViewModel.perform(action: .faceObservationDetected(faceObservationModel))
         } else {
             // Fallback on earlier versions
         }
@@ -343,7 +200,6 @@ extension FaceDetectorV2 {
         let faceQualityModel = FaceQualityModel(
             quality: result.faceCaptureQuality ?? 0.0
         )
-        self.faceDetectedState = .faceDetected
-        self.faceQualityState = .faceFound(faceQualityModel)
+        selfieViewModel.perform(action: .faceQualityObservationDetected(faceQualityModel))
     }
 }
