+++ Sources/SmileID/Classes/SelfieCapture/SelfieViewModelAction.swift
@@ -5,6 +5,9 @@ enum SelfieViewModelAction {
     case windowSizeDetected(CGRect)
 
     // Face Detection Actions
+    case faceObservationDetected(FaceGeometryModel)
+    case faceQualityObservationDetected(FaceQualityModel)
+    case selfieQualityObservationDetected(SelfieQualityModel)
     case updateUserInstruction(SelfieCaptureInstruction?)
     case activeLivenessCompleted
     case activeLivenessTimeout
